# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AMR uygulaması API",
    "version": "2.0.0",
    "description": "Kanzasset ↔ AMR sözleşmesi. Kimlik: X-API-Key, X-Timestamp, X-Signature (HMAC-SHA256: ts + yöntem + yol + ham gövde). POST için Idempotency-Key. Miktarlar tam sayı (mg, cent), fiyatlar ondalık dize. Rafineri tarafında mint / burn / token kavramı yoktur."
  },
  "paths": {
    "/v1/session/status": {
      "get": {
        "summary": "Oturum durumu: yayın açık mı, merkez bağlı mı",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "X-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "HMAC-SHA256(secret, ts + METHOD + path + body) hex"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "tradable",
                    "source_connected",
                    "ts"
                  ],
                  "properties": {
                    "status": {
                      "anyOf": [
                        {
                          "const": "OPEN",
                          "type": "string"
                        },
                        {
                          "const": "HALTED",
                          "type": "string"
                        },
                        {
                          "const": "MAINTENANCE",
                          "type": "string"
                        }
                      ]
                    },
                    "tradable": {
                      "type": "boolean"
                    },
                    "source_connected": {
                      "description": "merkez bağlantısı",
                      "type": "boolean"
                    },
                    "halt_reason": {
                      "type": "string"
                    },
                    "ts": {
                      "description": "ISO 8601 UTC",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/orders": {
      "post": {
        "summary": "Alış / satış emri (03, 04, 07, 08, 09). FOK: tümü ya hiç. Aynı client_order_id ile tekrar aynı cevabı alır.",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "X-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "HMAC-SHA256(secret, ts + METHOD + path + body) hex"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "client_order_id",
                  "side",
                  "qty_mg",
                  "ccy",
                  "quote_seq",
                  "limit_px",
                  "tif",
                  "time_limit_ms"
                ],
                "properties": {
                  "client_order_id": {
                    "description": "KZ tarafında tekil; aynı id ile tekrar gelen istek aynı cevabı alır",
                    "type": "string"
                  },
                  "side": {
                    "anyOf": [
                      {
                        "const": "BUY",
                        "type": "string"
                      },
                      {
                        "const": "SELL",
                        "type": "string"
                      }
                    ]
                  },
                  "qty_mg": {
                    "minimum": 1,
                    "type": "integer"
                  },
                  "ccy": {
                    "anyOf": [
                      {
                        "const": "USD",
                        "type": "string"
                      },
                      {
                        "const": "EUR",
                        "type": "string"
                      },
                      {
                        "const": "AED",
                        "type": "string"
                      }
                    ]
                  },
                  "quote_seq": {
                    "description": "emrin dayandığı fiyat tick'i",
                    "type": "integer"
                  },
                  "limit_px": {
                    "pattern": "^-?\\d+(\\.\\d+)?$",
                    "description": "ondalık dize, ör. \"142.00\"",
                    "type": "string"
                  },
                  "tif": {
                    "const": "FOK",
                    "type": "string"
                  },
                  "time_limit_ms": {
                    "minimum": 100,
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "FILLED / REJECTED / CANCELLED; FILLED cevabında account (bakiye bilgisi) ve alışta allocation_certificate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "order_id",
                    "client_order_id",
                    "status",
                    "side",
                    "qty_mg",
                    "ccy",
                    "quote_seq",
                    "limit_px",
                    "received_ts"
                  ],
                  "properties": {
                    "order_id": {
                      "type": "string"
                    },
                    "client_order_id": {
                      "type": "string"
                    },
                    "status": {
                      "anyOf": [
                        {
                          "const": "RECEIVED",
                          "type": "string"
                        },
                        {
                          "const": "CANCEL_REQUESTED",
                          "type": "string"
                        },
                        {
                          "const": "FILLED",
                          "type": "string"
                        },
                        {
                          "const": "REJECTED",
                          "type": "string"
                        },
                        {
                          "const": "CANCELLED",
                          "type": "string"
                        }
                      ]
                    },
                    "side": {
                      "anyOf": [
                        {
                          "const": "BUY",
                          "type": "string"
                        },
                        {
                          "const": "SELL",
                          "type": "string"
                        }
                      ]
                    },
                    "qty_mg": {
                      "type": "integer"
                    },
                    "ccy": {
                      "anyOf": [
                        {
                          "const": "USD",
                          "type": "string"
                        },
                        {
                          "const": "EUR",
                          "type": "string"
                        },
                        {
                          "const": "AED",
                          "type": "string"
                        }
                      ]
                    },
                    "quote_seq": {
                      "type": "integer"
                    },
                    "limit_px": {
                      "pattern": "^-?\\d+(\\.\\d+)?$",
                      "description": "ondalık dize, ör. \"142.00\"",
                      "type": "string"
                    },
                    "fill": {
                      "type": "object",
                      "required": [
                        "px",
                        "qty_mg",
                        "amount_cents",
                        "ccy",
                        "trade_ts"
                      ],
                      "properties": {
                        "px": {
                          "pattern": "^-?\\d+(\\.\\d+)?$",
                          "description": "ondalık dize, ör. \"142.00\"",
                          "type": "string"
                        },
                        "qty_mg": {
                          "type": "integer"
                        },
                        "amount_cents": {
                          "type": "integer"
                        },
                        "ccy": {
                          "anyOf": [
                            {
                              "const": "USD",
                              "type": "string"
                            },
                            {
                              "const": "EUR",
                              "type": "string"
                            },
                            {
                              "const": "AED",
                              "type": "string"
                            }
                          ]
                        },
                        "trade_ts": {
                          "description": "ISO 8601 UTC",
                          "type": "string"
                        }
                      }
                    },
                    "reject_reason": {
                      "anyOf": [
                        {
                          "const": "PRICE_OUTSIDE_LIMIT",
                          "type": "string"
                        },
                        {
                          "const": "STALE_QUOTE",
                          "type": "string"
                        },
                        {
                          "const": "TRADING_HALTED",
                          "type": "string"
                        },
                        {
                          "const": "CURRENT_ACCOUNT_LIMIT",
                          "type": "string"
                        },
                        {
                          "const": "DUPLICATE_ORDER",
                          "type": "string"
                        },
                        {
                          "const": "INVALID_QTY",
                          "type": "string"
                        },
                        {
                          "const": "INSUFFICIENT_CURRENT_ACCOUNT",
                          "type": "string"
                        },
                        {
                          "const": "INSUFFICIENT_VAULT",
                          "type": "string"
                        },
                        {
                          "const": "QUOTE_EXPIRED",
                          "type": "string"
                        },
                        {
                          "const": "INTERNAL_ERROR",
                          "type": "string"
                        }
                      ]
                    },
                    "allocation_certificate": {
                      "type": "object",
                      "required": [
                        "doc_id",
                        "url"
                      ],
                      "properties": {
                        "doc_id": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        }
                      }
                    },
                    "account": {
                      "type": "object",
                      "required": [
                        "seq",
                        "vault",
                        "current_account",
                        "status"
                      ],
                      "properties": {
                        "seq": {
                          "description": "hesap hareket numarası (fiyat seq'inden ayrı)",
                          "type": "integer"
                        },
                        "vault": {
                          "type": "object",
                          "required": [
                            "in_vault_mg",
                            "placing_mg",
                            "shipping_mg"
                          ],
                          "properties": {
                            "in_vault_mg": {
                              "description": "kasada",
                              "type": "integer"
                            },
                            "placing_mg": {
                              "description": "kasaya konuluyor (kabul edildi, en geç T+3)",
                              "type": "integer"
                            },
                            "shipping_mg": {
                              "description": "sevkiyatta (çıktı, henüz teslim edilmedi)",
                              "type": "integer"
                            }
                          }
                        },
                        "current_account": {
                          "type": "object",
                          "required": [
                            "gold_mg",
                            "money"
                          ],
                          "properties": {
                            "gold_mg": {
                              "description": "T: artı = aldık, kasaya konmadı; eksi = sattık, kasadan çıkmadı",
                              "type": "integer"
                            },
                            "money": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "required": [
                                  "ccy",
                                  "cents"
                                ],
                                "properties": {
                                  "ccy": {
                                    "anyOf": [
                                      {
                                        "const": "USD",
                                        "type": "string"
                                      },
                                      {
                                        "const": "EUR",
                                        "type": "string"
                                      },
                                      {
                                        "const": "AED",
                                        "type": "string"
                                      }
                                    ]
                                  },
                                  "cents": {
                                    "description": "eksi = Kanzasset borçlu, artı = rafineri borçlu",
                                    "type": "integer"
                                  }
                                }
                              }
                            }
                          }
                        },
                        "status": {
                          "anyOf": [
                            {
                              "const": "OK",
                              "type": "string"
                            },
                            {
                              "const": "RECONCILE",
                              "type": "string"
                            },
                            {
                              "const": "HALTED",
                              "type": "string"
                            }
                          ]
                        }
                      }
                    },
                    "received_ts": {
                      "description": "ISO 8601 UTC",
                      "type": "string"
                    },
                    "decided_ts": {
                      "description": "ISO 8601 UTC",
                      "type": "string"
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "status",
                          "ts"
                        ],
                        "properties": {
                          "status": {
                            "anyOf": [
                              {
                                "const": "RECEIVED",
                                "type": "string"
                              },
                              {
                                "const": "CANCEL_REQUESTED",
                                "type": "string"
                              },
                              {
                                "const": "FILLED",
                                "type": "string"
                              },
                              {
                                "const": "REJECTED",
                                "type": "string"
                              },
                              {
                                "const": "CANCELLED",
                                "type": "string"
                              }
                            ]
                          },
                          "ts": {
                            "description": "ISO 8601 UTC",
                            "type": "string"
                          },
                          "note": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "geçersiz emir"
          },
          "401": {
            "description": "kimlik"
          },
          "409": {
            "description": "client_order_id başka gövdeyle kullanıldı (DUPLICATE_ORDER)"
          }
        }
      }
    },
...
    "/v1/catalog": {
      "get": {
        "summary": "Rafinasyon ürün kataloğu (11): ürün · gramaj · ayar · tarife · üretim süresi. Değişince catalog.updated olayı gider.",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "X-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "HMAC-SHA256(secret, ts + METHOD + path + body) hex"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "version",
                    "items",
                    "updated_ts"
                  ],
                  "properties": {
                    "version": {
                      "type": "integer"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "item_id",
                          "name",
                          "weight_mg",
                          "fineness",
                          "unit_price_cents",
                          "ccy",
                          "lead_time_days",
                          "active"
                        ],
                        "properties": {
                          "item_id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "weight_mg": {
                            "minimum": 1,
                            "description": "ürünün saf gramajı",
                            "type": "integer"
                          },
                          "fineness": {
                            "description": "ayar, ör. 999.9",
                            "type": "string"
                          },
                          "unit_price_cents": {
                            "minimum": 0,
                            "description": "kalem başına tarife (işçilik)",
                            "type": "integer"
                          },
                          "ccy": {
                            "anyOf": [
                              {
                                "const": "USD",
                                "type": "string"
                              },
                              {
                                "const": "EUR",
                                "type": "string"
                              },
                              {
                                "const": "AED",
                                "type": "string"
                              }
                            ]
                          },
                          "lead_time_days": {
                            "minimum": 0,
                            "type": "integer"
                          },
                          "active": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "updated_ts": {
                      "description": "ISO 8601 UTC",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}